requirements.txt
jsx-lexer==2.0.1
Pygments==2.16.1
